Split renderers.py to multiple files
